Significantly changed the way riak_repl_leader works.
Riak_repl_leader has been rewritten as a gen_server which uses a
helper gen_leader process.  The gen_leader name includes
a hash of the candidates/workers to ensure that the list of
candidate nodes for any election is fixed.

The rewritten repl_leader includes extra stats about how many
objects are sent to the other cluster(s), how many were
forwarded to the leader and how many dropped.