Bug 1044: Rapid, continuous failure of replication TCP connections causes replication clients to drop all connections
Change MaxR for riak_repl_client_sup from 10 -> 100 in 10 seconds.
This prevents the self-inflicted DoS that the supervisor is designed
to do.  I think there are cases where we do want the child worker
to fail & not be restarted when there are enough failures quickly,
but there's a trade-off: 10 is too low, 50,000 is too high.  Hopefully
100 is just right.

Change MaxR 10 -> 100 and restart type from temporary -> transient for
the riak_repl_listener_sup.  Same "10 is too low" reasoning for MaxR.
For the restart type, if a listener process fails/is crashed by init,
then a 'temporary' proc won't be restarted and thus the server's TCP
service isn't available.  When using the 'transient' type, the TCP
service will be available even after 5 minutes of this kind of
punishment (a variation of Dan's original recipe, see the ticket):

lists:foldl(fun(_,_) -> riak_repl_listener:close_all_connections(),
                        timer:sleep(500)
            end, [], lists:seq(1,5*120)).