BZ1074 Don't close the repl server socket if the election is still going
Previous to this commit if, for whatever reason, the gen_leader election
was taking less than a small amount of time and a replication connection
was established the repl_tcp_server would instantly close the
connection. This in turn led to the connecting client exiting and
hitting the supervisor's max restarts and having the supervisor restart.

This change makes the code wait up to 60 seconds for the result of an
election to to become known, checked in 5 second intervals. Until we
know what the status of the election is, we do not consume any data off
the socket (we put the socket in passive mode). If the timeout is hit,
we exit as before, but if the election completes we either start doing
replication or refer the client to the leader. A new starting state was
added to the FSM, send_peerinfo which is where the FSM waits until it
knows who the leader is. Once we know who the leader is we put the
socket in {active, once} mode and continue as before.