Randomize the partition list on fullsync start
bz1273

Instead of keeping fullsync state in the cluster somehow, instead when
the current state of a fullsync is lost (node crashes, new leader is
elected), just shuffle the partition list instead. This at least ensures
that partitions already fullsynced by the aborted fullsync won't be
synced again so soon.