Fix another process leak; add put pool size to client stats
The put_fsms in the repl_do_put function were not blocking, so while the
worker pool was managing the # of workers, the # of put processes was
not being controlled. Now, the worker waits for the put_fsm to reply
before it returns and handles the next piece of work.