Move lambda into proper function to avoid distributed lambda problems
This function was purposefully created because if you use a lambda
then things will go wrong when the MD5 of the module changes. I.e. if
the lambda is shipped to another node with a different version of
`riak_repl_fullsync_helper`, even if the code inside the lambda is the
same, then a `badfun` error will occur since the MD5s of the modules
are not the same.

See http://www.javalimit.com/2010/05/passing-funs-to-other-erlang-nodes.html