Balance replication clients across all non-leader nodes
This change attempts to balance the load imposed by replication on the
leader node, by allowing the clients to be run on the non-leader nodes,
but the leader coordinates the distribution of clients across the
cluster.

For cases where there is only one alive node in the cluster, the old
behaviour is retained, and all clients run on the leader. In any other
case, clients are spread as evenly as possible across the cluster. When
new nodes are added or leave, the clients are rebalanced to try to
preserve fairness, while also minimizing churn.