Fix repl so it can be disabled in default_bucket_props
The code to delete the repl property before adding it back to the
default bucket properties was a holdover from when it was deleting the
repl postcommit hook and then re-adding it, in case it had changed.

This is actually harmful now, because it means you CANNOT disable the
realtime repl hook globally, if you wish (and enable it only on specific
buckets). The code has therefore been removed.