Implement vnode folding with a bloom filter to exchange differences
Because we traverse the vnode in disk-order while folding, if we create
a bloom filter of all the changed keys and then fold over the vnode, we
read keys off the disk in disk order, and don't pay the penalty for
random disk access. This does double the number of folds done in a
fullsync, but seems to be worth the price.