Avoid dropping realtime messages right after election
The default mailbox size of the elected leader is 'undefined' until a
mailbox check message comes in. This means that all messages received
before the first mailbox check will be dropped. This patch changes the
behaviour to reset the elected leader's known mailbox size to 0 on every
election. Additionally, the record defaults the mailbox size to 0 now,
rather than 'undefined'.