Use subnet masking to figure out the right cluster IPs
Assuming all your nodes are on the same subnet, and you bind the cluster
manager to 0.0.0.0, another cluster connecting will send the ip it
*thinks* it is connecting to. We can use that information to detect NAT,
as well as figure out what IP/subnet they connected to and find the IPs
on that same subnet for all the other nodes in the cluster.

If you bind to a specific IP, the autodetection will not run and the
specific bind IP will be returned instead. NAT is detected but currently
not supported.