Starting and running fssources
If a source dies unexpectedly, it's put at the back of the queue.  If
everything is perfect, then partitions are synced in the order they are
sorted by.  If something goes wrong, such as a node having no slots
availble for a new fssoruce, or an fssource dies unexpectedly, the sort
becomes a suggestion; a very loose often ignored suggestion :/