Improve realtime migration when riak_repl is stopping
These changes are intended to avoid dropping realtime messages that
happen or are pending during riak_repl stopping. The key changes are
that we try to send realtime messages to the rtq_proxy first, which
avoids a race with the process dictionary. We also explicitly stop() the
repl2_rtq so that we can gurantee terminate() completes. To do that, we
needed to make it a transient child of its supervisor, so it wouldn't
immediately restart.