send a clear message to ALL nodes
This is to prevent some nodes from getting stuck in a dirty state. For
example, if the coordinator isn't running, but a rt_dirty is set, the
coordinator won't know about the dirty state. It's still useful to have
the coordinator maintain a list of dirty nodes, as they appear in the
fscoord stats.