Fixed getting leader node in case of cluster_mgr crash
This swaps the start order of the cluster_mgr and riak_repl2_leader.
The problem solved is the case where the cluster_mgr would crash. If an
election was a long time off, the cluster_mgr would sit in a non-leadered
state indefinately.

By having the cluster_mgr have an explicit dependance on riak_repl2_leader,
(ie, start some time after) and asking about the leader, the above scenario
does not occur.

Some other tests needed to be altered to account for the new explicit
dependancy.