Added teardown to eunit to avoid unneeded output
If a meck is not unloaded before the test process dies, and cover is
enabled, a large amount of output is printed. Cover is unable to find the
source file for the meck'ed module, and ends up printing the module raw
binary. Because this is not an error, it's best to simply silence it.