Fixed function clause error when migrating queue
Discovered during a riak test, where shutting down a node would hang
forever chewing up a significant amount of cpu.

The queue migration delivery fun expected a tuple of size 3, but was given
a tuple of size 4 (due to the meta data). Fixed the clause and added a down
cast to push a size 3 to the remote rtq in case it's an older version.