Refactored stateful
Instead of a pure meck, using an actual module. All it does is start a
process and use the process dictionary. This change is due to a race
condition when using the meck. If two processes attempted to update a
stateful key, meck would exit due to code load concurrentcy.