Attempts to encode a datetime get converted to a string
While a fullsync is running, the fscoordinator stats have a datetime
stating when it was started. This does not play well with json encoding.
This uses the jsonify stats to find any datetimes (because general is
better than specific) and convert them to a date string (same rfc that
is used for now() tuples).