Ensure we re-connect to leader.
Prevent connections to non-leader nodes, when the sink leader changes,
because that's where all proxy_get requests were originating.  This
prevents a non-leader node from being the primary connection for the
proxy_get requests.

In addition, switch a bunch of registration calls to be synchronous, to
prevent races during initialization.