Run notify funs in order of registration, not in reverse order
This fixes an issue with the proxy being restarted because the proxy
notify hook is registered *after* the cluster_mgr hook. This means that
without this patch, the proxy hook runs before the cluster_mgr hook and
when the block provider asks the cluster_mgr for the leader, there's a
race condition where it can get the *previous* leader back.