Removed receive timeout that was crashing vnode.
If the receive timeout was reached, but the message it was waiting for
did eventually arrive, message would be put through the vnode's
handle_info, which does not have a clause for it. The usual vnode handlers
for the message (a response from another vnode) are not timed out, so
this should be a safe move.