Added check to not send request to self().
This prevents a vnode from sending a local lookup to itself, instead
falling back to the proxy_get. This was causing a vnode to become
dead-locked as it would send a message to itself then wait for a message
to be sent from itself; if it had to get any other messages in between, it
was boned.