Filter consistent objects during fullsync replication
Replication of consistent objects is not supported in Riak 2.0. Add
code to check the bucket properties of an object during fullsync and
prevent replication if the consistent property is set. It is not
necessary to filter realtime write at this time because the postcommit
hook used for realtime is not invoked on the consistent code path.