Include a bucket type hash in the message encoding for fullsync
Certain properties of typed buckets need to be identical on the source
and sink as a condition for data replication between buckets using
those bucket types.. For fullsync replication encode a hash of these
bucket type properties in the messages sent from the source.  The sink
side performs a comparison of a hash of its local bucket type
properties for the corresponding bucket type and only writes the data
if the hashes are identical.