Fix case where AAE fullsync may never terminate
Using AAE fullsync when a bucket has differing n_val between the
source and sink clusters can cause the fullsync process to go into an
infinite loop. When this situation occurs the riak_repl_aae_source
worker is terminated for the partition and riak_repl2_fssource
downgrades the strategy to keylist and establishes a new connection
via the connection manager.

What should happen next is that the fullsync for the partition is
retried using the keylist strategy. In actuality what happens is that
AAE fullsync is again selected as the strategy and the AAE tree
comparison begins again and then again fails. This process repeats ad
infinitum. This can be demonstrated by running the
repl_aae_fullsync_custom_n riak_test.

The reason this happens is that riak_repl2_fssource checks the value
of the fullsync_strategy environment variable every time
decide_our_caps is called. In the case of AAE fullsync, where
fullsync_strategy must be set for it to be enabled this has the effect
of overriding the attempt to downgrade to the keylist strategy and
results in the fullsync never terminating.

The fix to this problem is to change riak_repl2_fssource to only check
the fullsync_strategy environment variable during initialization and
to treat the strategy field of the state record as the canonical value
for the current requested strategy instead of repeatedly going back to
the application environment to obtain this value. This prevents the
downgraded strategy value from being overwritten and allow the
fullsync to complete even in situations of differing n_val values
among clusters.