Check for consumers before pushing data onto the realtime queue
Sink clusters that do not serve as source cluster for other clusters
have data build up on the realtime queue when using the default
settings. There are two factors that contribute to this: the fact that
`realtime_cascades` has a default value of `always` and the fact that
there is no check to ensure there are any realtime queue consumers
before data is pushed onto the queue. If `realtime_cascades` is set to
`always` data is always pushed onto the queue regardless of whether
there are any consumers for it or not.  This change resolves that
issues by adding a check to the handling of data pushes in
riak_repl2_rtq. If the list of consumers is empty the data is not
placed on the queue and if the list is populated the behavior remains
the same.