Optimize AAE replication.
Implement a prototype optimization for active anti-entropy replication
which front-loads the first few levels of bucket retrieval through one
TCP request, instead of 1025 roundtrips.  Futher optimizations include
increasing the number of levels which are prefetched.

This is to generally vet the changeset, and explore possible ways of
mitigating the protocol fallback and backwards compatability.