Use valid API for starting strategy modules in riak_repl_tcp_server
Commit 0b6acd2d introduced a change to the API for starting fullsync
strategy modules. Specifically, a parameter was added to the
start_link function for each fullsync strategy module. Change the call
in riak_repl_tcp_server to start the strategy module to include the
additional required parameter. The extra parameter indicates the
replication protocol level supported and is set to {1,0} in this case
since the code is only executed for legacy connections.