Add experimental schema.
These settings were extracted from the replication code, but have not
been vetted or necessarily named properly. Many of them are slated to be
removed or changed for future versions of riak_repl. To promote a
setting to the main schema, one should write documentation and tests to
cover the setting.