Issue Riak-1338
Cleaned up the care and feeding of the fake_sink test server to use a dynamically allocated port.  Also attempt to make sure it gets shut down cleanly so that the reported address-in-use error would be less likely, though the port could be in use otherwise so a dynamic one is superior all around.

Revised the riak_repl2_rtsource_conn eunit test to throw an error on failure, which is what eunit is looking for to determine pass/fail.

Removed a block of "receive" code at the end of the riak_repl2_rtsource_conn test that always returned a timeout error, as nothing was sending it a message at that point.

Refactored a teensy bit of test code so the names make sense.