Fixed send_after call, and removed un-needed calls.
Based on feedback, resetting the keep-alive timer on each socket message
may not be strictly needed. We just need some data every 60 seconds or so
to ensure a node is alive (rather than the tcp timeout which could be
much longer). An interval is still not used (reasons in previous commit
still stand), so instead it's just re-creating the timer on each keepalive.