Removed keepalive_timer from state.
We never cancel it, nor refer to the timer in the code, ever. One could
flood the process by manually sending 'keepalive' to it, but if someone is
doing that, you've got bigger problems.

This means that now the timer is like a send interval that only intervals
if the process is not stuck, and does not add a timer on proxy get
requests.