Critical replication bug fix - table comparison (#768)
* Critical replication bug fix - table comparison

Previously, if a TS table did not exist in the sink cluster, the arrival
of data via replication destined for that table would cause a recurring
process failure until the node was restarted. All realtime replication
through that node would stop working.

This was a careless mistake on my part involving returning an error atom
from the code that retrieves metadata from the intended table.

`get_identity_hashes` is invoked in two places; once in the same module,
once in `riak_repl_bucket_type_util`. The latter expects a list as a
return value and blindly runs comparisons against it; those comparisons
fail badly when an atom is returned.

With this fix, logging on the sink clearly indicates that the code is
behaving properly.

```
2016-12-16 09:58:53.494 [warning] <0.4866.0>@riak_repl2_ts:get_identity_hashes:97 DDL module for <<"foo">> does not exist or is not compiled with identity hashes
2016-12-16 09:59:03.496 [error] <0.4866.0>@riak_repl2_rtsink_conn:handle_info:237 drops due to missing or mismatched type <<"foo">>: 1
2016-12-16 09:59:03.496 [error] <0.4866.0>@riak_repl2_rtsink_conn:handle_info:243 total bucket type drop count: 1
```