gh774 fix: ensure ebloom is passed a binary
As per the issue https://github.com/basho/riak_repl/issues/774 in some
instances the Bucket given to ebloom was a 2-tuple (typed bucket) not a
binary, and so a `badarg` was thrown. Though this does not stop repl
from working, it leads to crashes and re-tries, is broken, and polutes
this log. This commit fixes by ensuring a binary is passed to
ebloom:insert/2 and ebloom:contains/2