Add an API to read an put an object on the realtime repl queue
The realtime repl queue can drop objects. It may be possible also to
detect inter-cluster divergence by external mechanisms (for example
writing last-modified as an index and comparing query results between
clusters.) An in progress feature allows rapid cluster contents
divergence checking with tic-tac-tree based AAE. Any of these use
cases needs a way to prompt replication of an object. This feature
provides that.

The API accepts a bucket/key pair, and the many of the same options as
a GET. Riak then fetches the object, and enqueues it on the repl
queue. Returning ok or an error.

NOTE: this work was originally merged to riak_kv, but that was the
wrong place for it. This is re-working and move of the code in
https://github.com/martinsumner/riak_kv/pull/6