Inermittent unit test failures
When setting up the sink for a test it might receive a heartbeat before the transport has been setup - soingore the heartbeat in this case.  This only seems to be tiggered on Travis CI.

In riak_repl2_rtq tests intermittently the Seq2 might race with Seq1?  Other such events ahve a sleep in-between of 1200, so added here.